requirements.txt
pytest
pyyaml
jinja2
pytest-cov
boto
pylint
pyrx-ats
